Rename constraints._mixin to constraints._support
